refactor(avatars): extract avatar image mechanics into HasAvatarImage trait
Moves the generic avatar plumbing off User into a reusable
App\Models\Traits\HasAvatarImage: the single-file media collection, the
thumb/medium conversions, getAvatarMedia(), and the getAvatar() URL
bundle. The AVATAR_COLLECTION / AVATAR_MIME_TYPES constants live on the
trait (PHP 8.4 trait constants), so User::AVATAR_COLLECTION still
resolves and external references are unchanged.

Collection registration is exposed as registerAvatarMediaCollection() /
registerAvatarMediaConversions() rather than owning the Spatie hooks, so
it composes with a model's other collections; avatarCollectionName() /
avatarMimeTypes() are overridable. A future publication icon can reuse
the mechanics with use HasAvatarImage.

Deliberately NOT moved: avatar moderation (the uploadAvatar gate,
upload-blocked flag, report/snapshot/audit system) stays on User — a
separate, user-specific concern, not part of 'has an avatar image'.

Backend suite 656 green, lint clean.
